fix(ci): move helpers to a separate file
Avoid capturing the script instance in the closures
